+++ frontend/src/manage/notifications/notificationsByOrganisationView.js
@@ -2,8 +2,9 @@ import React, {Component} from 'react';
 import {Col, Container,DropdownButton, Dropdown, Row} from "react-bootstrap";
 import {gql} from "apollo-boost";
 import QueryHandler from "../../utils/queryHandler";
-import MailContactsEditMutation from "./mailContactsEditMutation";
+import MailContactsEditMutation from "./contacts/mailContactsEditMutation";
 import NotificationSubscriptionsEditMutation from "./notificationSubscriptionsEditMutation";
+import FirebaseMessageContactsEditMutation  from "./contacts/firebaseMessageContactsEditMutation";
 import EditableTable from "../../components/editableTable";
 
 const NOTIFICATION_CONFIG_BY_ORGANISATION_ID = gql`
@@ -17,6 +18,9 @@ query organisationById($id: ID!) {
         __typename
         ... on MailContact {
           mailAddress
+        }
+        ... on FirebaseMessageContact {
+          mail
         }
       }
       subscriptions {
@@ -56,6 +60,12 @@ class NotificationsByOrganisationView extends Component {
             .contacts
             .filter(contact => contact.__typename == "MailContact");
 
+            let firebaseMessageContacts = data
+            .organisationById
+            .notificationConfig
+            .contacts
+            .filter(contact => contact.__typename == "FirebaseMessageContact");
+
             return (
                 <Container>
                   <Row className={"row-header"}>
@@ -67,6 +77,13 @@ class NotificationsByOrganisationView extends Component {
                           organisationId={this.props.match.params.id}
                           contacts={mailContacts}
                       />
+
+                      <h2>Firebase Push</h2>
+                      <FirebaseMessageContactsEditMutation
+                          onContactsChanged={() => refetch()}
+                          organisationId={this.props.match.params.id}
+                          contacts={firebaseMessageContacts}
+                        />
 
                     </Col>
                   </Row>