+++ src/main/java/com/alarmcontrol/server/data/graphql/InterfaceConfiguration.java
@@ -1,5 +1,6 @@
 package com.alarmcontrol.server.data.graphql;
 
+import com.alarmcontrol.server.notifications.messaging.firebasepush.FirebaseMessageContact;
 import com.alarmcontrol.server.aao.config.AaoRule;
 import com.alarmcontrol.server.notifications.messaging.mail.MailContact;
 import com.alarmcontrol.server.notifications.usecases.alertcreated.AlertCreatedNotificationConfig;
@@ -14,9 +15,9 @@ public class InterfaceConfiguration {
   public SchemaParserDictionary getSchemaParser() {
     SchemaParserDictionary dictionary = new SchemaParserDictionary();
     dictionary.add(MailContact.class);
+    dictionary.add(FirebaseMessageContact.class);
     dictionary.add(AlertCreatedNotificationConfig.class);
     dictionary.add(AaoRule.class);
-    //dictionary.add(BetweenTimeRangeRule.class);
     return dictionary;
   }
 }