+++ src/main/java/com/alarmcontrol/server/notifications/messaging/mail/MailMessageService.java
@@ -1,7 +1,12 @@
 package com.alarmcontrol.server.notifications.messaging.mail;
 
+import com.alarmcontrol.server.notifications.core.messaging.AbstractMessageService;
 import com.alarmcontrol.server.notifications.core.messaging.Message;
+import com.alarmcontrol.server.notifications.core.messaging.MessageService;
+import java.util.List;
 import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -9,7 +14,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
 @Service
-public class MailMessageService {
+public class MailMessageService extends AbstractMessageService<MailContact> {
+
+  private Logger logger = LoggerFactory.getLogger(MailMessageService.class);
+
   @Value("${spring.mail.host:}")
   private String host;
 
@@ -28,10 +36,13 @@ public class MailMessageService {
   private JavaMailSender javaMailSender;
 
   public MailMessageService(JavaMailSender javaMailSender) {
+    super(MailContact.class);
     this.javaMailSender = javaMailSender;
   }
 
-  public void send(String mailaddress, Message message){
+  @Override
+  protected void sendInternal(List<MailContact> contacts, Message message) {
+    Assert.notNull(contacts, "contacts is null");
     Assert.notNull(message, "message is null");
 
     if(StringUtils.isBlank(host)){
@@ -42,7 +53,7 @@ public class MailMessageService {
       throw new RuntimeException("Port is not set!");
     }
 
-    if(getUseAuth()){
+    if(useAuthForMailserver()){
       if(StringUtils.isBlank(username)){
         throw new RuntimeException("Auth is enabled, but username is not set!");
       }
@@ -52,14 +63,12 @@ public class MailMessageService {
       }
     }
 
-    if(StringUtils.isBlank(mailaddress)){
-      throw new RuntimeException("Mailaddress is not set!");
+    for (MailContact contact : contacts) {
+      sendEmail(contact, message);
     }
-
-    sendEmail(mailaddress, message);
   }
 
-  private Boolean getUseAuth() {
+  private Boolean useAuthForMailserver() {
     try{
       return Boolean.parseBoolean(useAuth);
     }catch(Exception e) {
@@ -67,13 +76,31 @@ public class MailMessageService {
     }
   }
 
-  private void sendEmail(String mailaddress, Message message) {
-    SimpleMailMessage msg = new SimpleMailMessage();
-    msg.setTo(mailaddress);
+  private void sendEmail(MailContact contact, Message message) {
+    try{
+      if(StringUtils.isBlank(contact.getMailAddress())){
+        throw new RuntimeException("Mailaddress is not set!");
+      }
 
-    msg.setSubject(message.getSubject());
-    msg.setText(message.getBody());
+      logger.info("Start sending message '{}' via {} to {}",
+          message.getSubject(),
+          message.getClass().getSimpleName(),
+          contact.getMailAddress());
 
-    javaMailSender.send(msg);
+      SimpleMailMessage msg = new SimpleMailMessage();
+      msg.setTo(contact.getMailAddress());
+
+      msg.setSubject(message.getSubject());
+      msg.setText(message.getBody());
+
+      javaMailSender.send(msg);
+
+      logger.info("Following message sent to {} \n: {}",
+          contact.getMailAddress(),
+          message);
+
+    }catch (Exception e){
+      logger.error("Can not send message to {} \ndue to: {}\n {}", contact, e.getMessage(), message, e);
+    }
   }
 }